Use Playwright for component tests
This commit moves the repo away from using jest and instead using the
new component testing ability of Playwright. This will mount components
within the DOM of a real browser and have it behave as expected, which
is awesome since we can then drop the need for things like jsdom.